+++ Service/Warehouse/API/Warehouse/ModifyShapeIntersectionQuery.hpp
@@ -26,7 +26,9 @@ public:
         /// Prepared query constructs cursors.
         friend class ModifyShapeIntersectionQuery;
 
-        EMERGENCE_BIND_IMPLEMENTATION_INPLACE (sizeof (uintptr_t) * 21u);
+        /// Cursor implementation could copy Shape inside to be more cache coherent and Shape could contain doubles,
+        /// which are 8-byte long on all architectures. Therefore we use uint64_t as base size type.
+        EMERGENCE_BIND_IMPLEMENTATION_INPLACE (sizeof (uint64_t) * 21u);
 
         explicit Cursor (std::array<uint8_t, DATA_MAX_SIZE> *_data) noexcept;
     };
